split counters/columnshard.h, change schema opType
