+++ ydb/core/tx/columnshard/transactions/operators/schema.h
@@ -44,7 +44,20 @@ private:
     virtual void DoFinishProposeOnComplete(TColumnShard& /*owner*/, const TActorContext& /*ctx*/) override {
     }
     virtual TString DoGetOpType() const override {
-        return "Scheme";
+        switch (SchemaTxBody.TxBody_case()) {
+            case NKikimrTxColumnShard::TSchemaTxBody::kInitShard:
+                return "Scheme:InitShard";
+            case NKikimrTxColumnShard::TSchemaTxBody::kEnsureTables:
+                return "Scheme:EnsureTables";
+            case NKikimrTxColumnShard::TSchemaTxBody::kAlterTable:
+                return "Scheme:AlterTable";
+            case NKikimrTxColumnShard::TSchemaTxBody::kAlterStore:
+                return "Scheme:AlterStore";
+            case NKikimrTxColumnShard::TSchemaTxBody::kDropTable:
+                return "Scheme:DropTable";
+            case NKikimrTxColumnShard::TSchemaTxBody::TXBODY_NOT_SET:
+                return "Scheme:TXBODY_NOT_SET";
+        }
     }
     virtual bool DoIsAsync() const override {
         return WaitPathIdsToErase.size();
